+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Minor fixes/improvements
 - Single source of truth for default settings
+- Add support for search pages
 
 ## Version 1.0